@@ -12,8 +12,8 @@
         "fideloper/proxy": "^4.2",
         "fruitcake/laravel-cors": "^2.0",
         "guzzlehttp/guzzle": "^6.3",
-        "laravel/framework": "^7.24",
-        "laravel/passport": "^9.3",
+        "laravel/framework": "^8.0",
+        "laravel/passport": "^10",
         "laravel/tinker": "^2.0"
     },
     "require-dev": {
@@ -22,7 +22,7 @@
         "fzaninotto/faker": "^1.9.1",
         "mockery/mockery": "^1.3.1",
         "nunomaduro/collision": "^4.1",
-        "phpunit/phpunit": "^8.5"
+        "phpunit/phpunit": "^9"
     },
     "config": {
         "optimize-autoloader": true,
