@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\Api\Crm\CrmCompanyTypeController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -12,20 +13,12 @@ use Illuminate\Support\Facades\Route;
 | is assigned the "api" middleware group. Enjoy building your API!
 |
 */
-Route::namespace('Api\Crm')
-     ->group(function () {
-         Route::group(['prefix' => 'crm', 'middleware' => 'auth:api'], function () {
-             Route::group(['prefix' => 'type'], function () {
-                 Route::get('/', 'CrmCompanyTypeController@index')
-                      ->name('crm_type_index');
-                 Route::get('/{id}', 'CrmCompanyTypeController@show')
-                      ->name('crm_type_show');
-                 Route::post('/', 'CrmCompanyTypeController@store')
-                      ->name('crm_type_store');
-                 Route::put('/{id}', 'CrmCompanyTypeController@update')
-                      ->name('crm_type_update');
-                 Route::delete('/{id}', 'CrmCompanyTypeController@destroy')
-                      ->name('crm_type_delete');
-             });
-         });
-     });
+Route::group(['prefix' => 'crm', 'middleware' => 'auth:api'], function () {
+    Route::group(['prefix' => 'type'], function () {
+        Route::get('/', [CrmCompanyTypeController::class, 'index'])->name('crm_type_index');
+        Route::get('/{id}', [CrmCompanyTypeController::class, 'show'])->name('crm_type_show');
+        Route::post('/', [CrmCompanyTypeController::class, 'store'])->name('crm_type_store');
+        Route::put('/{id}', [CrmCompanyTypeController::class, 'update'])->name('crm_type_update');
+        Route::delete('/{id}', [CrmCompanyTypeController::class, 'destroy'])->name('crm_type_delete');
+    });
+});
