@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 /**
@@ -24,5 +25,5 @@ use Illuminate\Database\Eloquent\Model;
  */
 class UserSettings extends Model
 {
-    //
+    use HasFactory;
 }
